requirements.txt
flask
faiss-cpu
openai
gunicorn
